requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
youtube-transcript-api>=0.6.0
google-api-python-client>=2.100.0
